chore: architectural cleanups
