+++ .clang-format
@@ -1,0 +1,6 @@
+---
+BasedOnStyle: Google
+Language: Cpp
+IndentWidth: 4
+ColumnLimit: 100
+...
